--- src/pages/address/components/content-address/index.tsx
@@ -1,8 +1,0 @@
-const ContentAddress: React.FC = () => {
-  return (
-    <div>
-      <div></div>
-    </div>
-  );
-};
-export default ContentAddress;
