+++ src/pages/index.ts
@@ -5,4 +5,3 @@ export { default as Main } from './main';
 export { default as MyPage } from './mypage';
 export { default as Onboarding } from './onboarding';
 export { default as Products } from './products';
-export { default as Address } from './address';
